Challenge 2.0 front+diseño

+++ fullStack/database/models/Movements.js
@@ -15,7 +15,7 @@ let cols = {
     },
     amount:{
 
-        type: dataTypes.INTEGER
+        type: dataTypes.DECIMAL
     },
     date: {
         type:dataTypes.STRING
